refactor(Car, CarTest): Car 을 model 이하로 이동해라
